refactor(find-largest-number): Use type "size_t" instead of "int" because it's better for array index

+++ src/03-find-largest-number/index.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int get_largest_number(int numbers[], int numbers_length) {
+int get_largest_number(int numbers[], size_t numbers_length) {
   int largest_number = numbers[0];
 
   for (size_t i = 1; i < numbers_length; i++) {
@@ -16,7 +16,7 @@ int get_largest_number(int numbers[], int numbers_length) {
 
 int main() {
   int numbers[] = {4, 1, 7, 9, 2, 6, 8, 5, 3, 10};
-  int numbers_length = sizeof numbers / sizeof numbers[0];
+  size_t numbers_length = sizeof numbers / sizeof numbers[0];
 
   printf("The largest number is: %d\n", get_largest_number(numbers, numbers_length));
